backend/requirements.txt
altgraph==0.10.2
asn1crypto==0.24.0
backports.functools-lru-cache==1.6.1
bdist-mpkg==0.5.0
BeautifulSoup==3.2.2
beautifulsoup4==4.8.2
bonjour-py==0.3
bs4==0.0.1
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
enum34==1.1.6
excel==1.0.0
future==0.17.1
google-search-results-serpwow==1.1.8
idna==2.9
ipaddress==1.0.22
lxml==4.5.0
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
numpy==1.16.6
pandas==0.24.2
py2app==0.7.3
pycparser==2.19
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==2.8.1
pytz==2013.7
requests==2.23.0
scipy==0.13.0b1
six==1.12.0
soupsieve==1.9.5
urllib3==1.25.8
vboxapi==1.0
xattr==0.6.4
xlrd==1.2.0
